Fixes to importer old APSIM to Next Gen
