+++ ApsimNG/Presenters/CLEM/ModelDetailsWrapperPresenter.cs
@@ -10,7 +10,6 @@
     using APSIM.Shared.Utilities;
     using Commands;
     using EventArguments;
-    using Importer;
     using Interfaces;
     using Models;
     using Models.CLEM;
