+++ ApsimNG/Presenters/CLEM/PropertyTreeTablePresenter.cs
@@ -11,7 +11,6 @@ namespace UserInterface.Presenters
     using APSIM.Shared.Utilities;
     using Commands;
     using EventArguments;
-    using Importer;
     using Interfaces;
     using Models;
     using Models.Core;
