+++ ApsimNG/Presenters/ExplorerPresenter.cs
@@ -434,15 +434,12 @@
         /// <summary>
         /// Adds a model to a parent model.
         /// </summary>
-        /// <param name="st">The string representation (JSON or XML) of a model.</param>
-        /// <param name="parentPath">Path to the parent</param>
-        public void Add(string st, string parentPath)
+        /// <param name="parentPath">Path to the parent.</param>
+        /// <param name="modelToAdd">The model to add to the tree.</param>
+        public void AddChildToTree(string parentPath, IModel modelToAdd)
         {
-            IModel model = FileFormat.ReadFromString<IModel>(st, out List<Exception> errors);
-            if (errors != null && errors.Count > 0)
-                throw errors[0];
-            AddModelCommand command = new AddModelCommand(parentPath, model, view, this);
-            CommandHistory.Add(command, true);
+            var nodeDescription = GetNodeDescription(modelToAdd);
+            view.Tree.AddChild(parentPath, nodeDescription);
         }
 
         /// <summary>
@@ -450,10 +447,9 @@
         /// </summary>
         /// <param name="child">The string representation (JSON or XML) of a model.</param>
         /// <param name="parentPath">Path to the parent</param>
-        public void Add(IModel child, string parentPath)
+        public void DeleteFromTree(string pathToNodeToDelete)
         {
-            AddModelCommand command = new AddModelCommand(parentPath, child, view, this);
-            CommandHistory.Add(command, true);
+            view.Tree.Delete(pathToNodeToDelete);
         }
 
         /// <summary>Deletes the specified model.</summary>
@@ -901,7 +897,10 @@
                     ICommand cmd = null;
                     if (e.Copied)
                     {
-                        this.Add(modelString, toParentPath);
+                        var command = new AddModelCommand(pathOfParent: toParentPath,
+                                                          textToAdd: modelString,
+                                                          this);
+                        CommandHistory.Add(command, true);
                     }
                     else if (e.Moved)
                     {
