+++ Tests/UnitTests/ApsimNG/Views/GraphViewTests.cs
@@ -17,6 +17,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using UnitTests.ApsimNG.Utilities;
+using UserInterface.Commands;
 using UserInterface.Presenters;
 using UserInterface.Views;
 using Utility;
@@ -110,17 +111,27 @@ namespace UnitTests.ApsimNG.Views
             IModel paddock = Apsim.Find(sims, typeof(Zone));
             Folder graphs = new Folder();
             graphs.Name = "Graphs";
-            explorer.Add(graphs, Apsim.FullPath(paddock));
+
+            var command = new AddModelCommand(pathOfParent: Apsim.FullPath(paddock),
+                                              graphs,
+                                              explorer);
+            explorer.CommandHistory.Add(command, true);
 
             // Add an empty graph to the folder.
             Models.Graph.Graph graph = new Models.Graph.Graph();
             graph.Name = "Graph";
-            explorer.Add(graph, Apsim.FullPath(graphs));
+            command = new AddModelCommand(pathOfParent: Apsim.FullPath(graphs),
+                                          graph,
+                                          explorer);
+            explorer.CommandHistory.Add(command, true);
 
             // Add an empty series to the graph.
             Models.Graph.Series series = new Models.Graph.Series();
             series.Name = "Series";
-            explorer.Add(series, Apsim.FullPath(graph));
+            command = new AddModelCommand(pathOfParent: Apsim.FullPath(graph),
+                                          series,
+                                          explorer);
+            explorer.CommandHistory.Add(command, true);
 
             // click on the series node.
             explorer.SelectNode(Apsim.FullPath(series));
